requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.5
django-allauth==0.52.0
django-cart==1.0.4
django-crispy-forms==1.14.0
django-filter==22.1
django-import-export==3.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
filters==1.3.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==22.0
Pillow==9.4.0
pkg_resources==0.0.0
psycopg2-binary==2.9.5
pycosat==0.6.3
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tablib==3.3.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
